refactor: move functions dealing with workspace to workspaceUtils
