+++ packages/core/src/codewhisperer/commands/startSecurityScan.ts
@@ -50,6 +50,7 @@ import {
 import { SecurityIssueTreeViewProvider } from '../service/securityIssueTreeViewProvider'
 import { ChatSessionManager } from '../../amazonqScan/chat/storages/chatSession'
 import { TelemetryHelper } from '../util/telemetryHelper'
+import { getWorkspacePaths } from '../../shared/utilities/workspaceUtils'
 
 const localize = nls.loadMessageBundle()
 export const stopScanButton = localize('aws.codewhisperer.stopscan', 'Stop Scan')
@@ -156,7 +157,7 @@ export async function startSecurityScan(
             })
         }
         const zipMetadata = await zipUtil.generateZip(editor?.document.uri, scope)
-        const projectPaths = zipUtil.getProjectPaths()
+        const projectPaths = getWorkspacePaths()
 
         const contextTruncationStartTime = performance.now()
         codeScanTelemetryEntry.contextTruncationDuration = performance.now() - contextTruncationStartTime
