+++ packages/core/src/testE2E/codewhisperer/securityScan.test.ts
@@ -23,6 +23,7 @@ import { makeTemporaryToolkitFolder } from '../../shared/filesystemUtilities'
 import fs from '../../shared/fs/fs'
 import { ZipUtil } from '../../codewhisperer/util/zipUtil'
 import { randomUUID } from '../../shared/crypto'
+import { getWorkspacePaths } from '../../shared/utilities/workspaceUtils'
 
 const filePromptWithSecurityIssues = `from flask import app
 
@@ -92,7 +93,7 @@ describe('CodeWhisperer security scan', async function () {
         const zipUtil = new ZipUtil()
         const uri = editor.document.uri
 
-        const projectPaths = zipUtil.getProjectPaths()
+        const projectPaths = getWorkspacePaths()
         const scope = CodeWhispererConstants.CodeAnalysisScope.PROJECT
         const zipMetadata = await zipUtil.generateZip(uri, scope)
         const codeScanName = randomUUID()
